GoStruct: specify fields using type IDs
This allows us to defer the resolution of type IDs to concrete Go
types until later -- which is essential for handling self-referential
types and out of order custom types in the gob stream.

From reading https://golang.org/pkg/encoding/gob/, I first thought
that all custom types would be defined in terms of previously defined
type IDs, and so it would make sense to use the concrete types here.
However, real-world gob streams contain custom types that refer to
later types, so we need to handle this case.